requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-author==1.2.0
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-js-asset==2.0.0
django-taggit==3.1.0
idna==3.4
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
setuptools-git==1.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.4.0
